move vespa functions to vespa_app.py
