+++ visual-retrieval-colpali/backend/colpali.py
@@ -13,7 +13,6 @@ import matplotlib.cm as cm
 import re
 import io
 
-import json
 import time
 import backend.testquery as testquery
 
@@ -24,12 +23,10 @@ from vidore_benchmark.interpretability.torch_utils import (
     normalize_similarity_map_per_query_token,
 )
 from vidore_benchmark.interpretability.vit_configs import VIT_CONFIG
-from vespa.application import Vespa
-from vespa.io import VespaQueryResponse
 
 matplotlib.use("Agg")
-
-MAX_QUERY_TERMS = 64
+# Prepare the colormap once to avoid recomputation
+colormap = cm.get_cmap("viridis")
 
 COLPALI_GEMMA_MODEL_NAME = "vidore/colpaligemma-3b-pt-448-base"
 
@@ -62,54 +59,6 @@ def load_vit_config(model):
     return vit_config
 
 
-def save_figure(fig, filename: str = "similarity_map.png"):
-    try:
-        OUTPUT_DIR = Path(__file__).parent.parent / "output" / "sim_maps"
-        OUTPUT_DIR.mkdir(parents=True, exist_ok=True)
-        fig.savefig(
-            OUTPUT_DIR / filename,
-            bbox_inches="tight",
-            pad_inches=0,
-        )
-    except Exception as e:
-        print(f"Failed to save figure: {e}")
-
-
-def annotate_plot(ax, query, selected_token):
-    # Add the query text as a title over the image with opacity
-    ax.text(
-        0.5,
-        0.95,  # Adjust the position to be on the image (y=0.1 is 10% from the bottom)
-        query,
-        fontsize=18,
-        color="white",
-        ha="center",
-        va="center",
-        alpha=0.8,  # Set opacity (1 is fully opaque, 0 is fully transparent)
-        bbox=dict(
-            boxstyle="round,pad=0.5", fc="black", ec="none", lw=0, alpha=0.5
-        ),  # Add a semi-transparent background
-        transform=ax.transAxes,  # Ensure the coordinates are relative to the axes
-    )
-
-    # Add annotation with the selected token over the image with opacity
-    ax.text(
-        0.5,
-        0.05,  # Position towards the top of the image
-        f"Selected token: `{selected_token}`",
-        fontsize=18,
-        color="white",
-        ha="center",
-        va="center",
-        alpha=0.8,  # Set opacity for the text
-        bbox=dict(
-            boxstyle="round,pad=0.3", fc="black", ec="none", lw=0, alpha=0.5
-        ),  # Semi-transparent background
-        transform=ax.transAxes,  # Keep the coordinates relative to the axes
-    )
-    return ax
-
-
 def gen_similarity_maps(
     model: ColPali,
     processor: ColPaliProcessor,
@@ -139,11 +88,6 @@ def gen_similarity_maps(
         Tuple[int, str, str]: A tuple containing the image index, the selected token, and the base64-encoded image.
 
     """
-
-    start = time.perf_counter()
-
-    # Prepare the colormap once to avoid recomputation
-    colormap = cm.get_cmap("viridis")
 
     # Process images and store original images and sizes
     processed_images = []
@@ -336,154 +280,6 @@ def get_query_embeddings_and_token_map(
     return q_emb, token_to_idx
 
 
-def format_query_results(query, response, hits=5) -> dict:
-    query_time = response.json.get("timing", {}).get("searchtime", -1)
-    query_time = round(query_time, 2)
-    count = response.json.get("root", {}).get("fields", {}).get("totalCount", 0)
-    result_text = f"Query text: '{query}', query time {query_time}s, count={count}, top results:\n"
-    print(result_text)
-    return response.json
-
-
-async def query_vespa_default(
-    app: Vespa,
-    query: str,
-    q_emb: torch.Tensor,
-    hits: int = 3,
-    timeout: str = "10s",
-    **kwargs,
-) -> dict:
-    async with app.asyncio(connections=1, total_timeout=120) as session:
-        query_embedding = format_q_embs(q_emb)
-
-        start = time.perf_counter()
-        response: VespaQueryResponse = await session.query(
-            body={
-                "yql": "select id,title,url,blur_image,page_number,snippet,text,summaryfeatures from pdf_page where userQuery();",
-                "ranking": "default",
-                "query": query,
-                "timeout": timeout,
-                "hits": hits,
-                "input.query(qt)": query_embedding,
-                "presentation.timing": True,
-                **kwargs,
-            },
-        )
-        assert response.is_successful(), response.json
-        stop = time.perf_counter()
-        print(
-            f"Query time + data transfer took: {stop - start} s, vespa said searchtime was {response.json.get('timing', {}).get('searchtime', -1)} s"
-        )
-        open("response.json", "w").write(json.dumps(response.json))
-    return format_query_results(query, response)
-
-
-async def query_vespa_bm25(
-    app: Vespa,
-    query: str,
-    q_emb: torch.Tensor,
-    hits: int = 3,
-    timeout: str = "10s",
-    **kwargs,
-) -> dict:
-    async with app.asyncio(connections=1, total_timeout=120) as session:
-        query_embedding = format_q_embs(q_emb)
-
-        start = time.perf_counter()
-        response: VespaQueryResponse = await session.query(
-            body={
-                "yql": "select id,title,url,blur_image,page_number,snippet,text,summaryfeatures from pdf_page where userQuery();",
-                "ranking": "bm25",
-                "query": query,
-                "timeout": timeout,
-                "hits": hits,
-                "input.query(qt)": query_embedding,
-                "presentation.timing": True,
-                **kwargs,
-            },
-        )
-        assert response.is_successful(), response.json
-        stop = time.perf_counter()
-        print(
-            f"Query time + data transfer took: {stop - start} s, vespa said searchtime was {response.json.get('timing', {}).get('searchtime', -1)} s"
-        )
-    return format_query_results(query, response)
-
-
-def float_to_binary_embedding(float_query_embedding: dict) -> dict:
-    binary_query_embeddings = {}
-    for k, v in float_query_embedding.items():
-        binary_vector = (
-            np.packbits(np.where(np.array(v) > 0, 1, 0)).astype(np.int8).tolist()
-        )
-        binary_query_embeddings[k] = binary_vector
-        if len(binary_query_embeddings) >= MAX_QUERY_TERMS:
-            print(f"Warning: Query has more than {MAX_QUERY_TERMS} terms. Truncating.")
-            break
-    return binary_query_embeddings
-
-
-def create_nn_query_strings(
-    binary_query_embeddings: dict, target_hits_per_query_tensor: int = 20
-) -> Tuple[str, dict]:
-    # Query tensors for nearest neighbor calculations
-    nn_query_dict = {}
-    for i in range(len(binary_query_embeddings)):
-        nn_query_dict[f"input.query(rq{i})"] = binary_query_embeddings[i]
-    nn = " OR ".join(
-        [
-            f"({{targetHits:{target_hits_per_query_tensor}}}nearestNeighbor(embedding,rq{i}))"
-            for i in range(len(binary_query_embeddings))
-        ]
-    )
-    return nn, nn_query_dict
-
-
-def format_q_embs(q_embs: torch.Tensor) -> dict:
-    float_query_embedding = {k: v.tolist() for k, v in enumerate(q_embs)}
-    return float_query_embedding
-
-
-async def query_vespa_nearest_neighbor(
-    app: Vespa,
-    query: str,
-    q_emb: torch.Tensor,
-    target_hits_per_query_tensor: int = 20,
-    hits: int = 3,
-    timeout: str = "10s",
-    **kwargs,
-) -> dict:
-    # Hyperparameter for speed vs. accuracy
-    async with app.asyncio(connections=1, total_timeout=180) as session:
-        float_query_embedding = format_q_embs(q_emb)
-        binary_query_embeddings = float_to_binary_embedding(float_query_embedding)
-
-        # Mixed tensors for MaxSim calculations
-        query_tensors = {
-            "input.query(qtb)": binary_query_embeddings,
-            "input.query(qt)": float_query_embedding,
-        }
-        nn_string, nn_query_dict = create_nn_query_strings(
-            binary_query_embeddings, target_hits_per_query_tensor
-        )
-        query_tensors.update(nn_query_dict)
-        response: VespaQueryResponse = await session.query(
-            body={
-                **query_tensors,
-                "presentation.timing": True,
-                # if we use rank({nn_string}, userQuery()), dynamic summary doesn't work, see https://github.com/vespa-engine/vespa/issues/28704
-                "yql": f"select id,title,snippet,text,url,blur_image,page_number,summaryfeatures from pdf_page where {nn_string} or userQuery()",
-                "ranking.profile": "retrieval-and-rerank",
-                "timeout": timeout,
-                "hits": hits,
-                "query": query,
-                **kwargs,
-            },
-        )
-        assert response.is_successful(), response.json
-    return format_query_results(query, response)
-
-
 def is_special_token(token: str) -> bool:
     # Pattern for tokens that start with '<', numbers, whitespace, or single characters, or the string 'Question'
     # Will exclude these tokens from the similarity map generation
@@ -491,55 +287,6 @@ def is_special_token(token: str) -> bool:
     if (len(token) < 3) or pattern.match(token):
         return True
     return False
-
-async def get_full_image_from_vespa(
-    app: Vespa,
-    id: str) -> str:
-    async with app.asyncio(connections=1, total_timeout=120) as session:
-        start = time.perf_counter()
-        response: VespaQueryResponse = await session.query(
-            body={
-                "yql": f"select full_image from pdf_page where id contains \"{id}\"",
-                "ranking": "unranked",
-                "presentation.timing": True,
-            },
-        )
-        assert response.is_successful(), response.json
-        stop = time.perf_counter()
-        print(
-            f"Getting image from Vespa took: {stop - start} s, vespa said searchtime was {response.json.get('timing', {}).get('searchtime', -1)} s"
-        )
-    return response.json["root"]["children"][0]["fields"]["full_image"]
-
-async def get_result_from_query(
-    app: Vespa,
-    processor: ColPaliProcessor,
-    model: ColPali,
-    query: str,
-    q_embs: torch.Tensor,
-    token_to_idx: Dict[str, int],
-    ranking: str,
-) -> Dict[str, Any]:
-    # Get the query embeddings and token map
-    print(query)
-
-    print(token_to_idx)
-    if ranking == "nn+colpali":
-        result = await query_vespa_nearest_neighbor(app, query, q_embs)
-    elif ranking == "bm25+colpali":
-        result = await query_vespa_default(app, query, q_embs)
-    elif ranking == "bm25":
-        result = await query_vespa_bm25(app, query, q_embs)
-    else:
-        raise ValueError(f"Unsupported ranking: {ranking}")
-    # Print score, title id, and text of the results
-    for idx, child in enumerate(result["root"]["children"]):
-        print(
-            f"Result {idx+1}: {child['relevance']}, {child['fields']['title']}, {child['fields']['id']}"
-        )
-    for single_result in result["root"]["children"]:
-        print(single_result["fields"].keys())
-    return result
 
 
 def add_sim_maps_to_result(
@@ -582,36 +329,3 @@ def add_sim_maps_to_result(
     #     for token, sim_mapb64 in sim_map_dict.items():
     #         single_result["fields"][f"sim_map_{token}"] = sim_mapb64
     return result
-
-
-if __name__ == "__main__":
-    model, processor = load_model()
-    vit_config = load_vit_config(model)
-    query = "How many percent of source water is fresh water?"
-    image_filepath = (
-        Path(__file__).parent.parent
-        / "static"
-        / "assets"
-        / "ConocoPhillips Sustainability Highlights - Nature (24-0976).png"
-    )
-    q_embs, token_to_idx = get_query_embeddings_and_token_map(
-        processor,
-        model,
-        query,
-    )
-    figs_images = gen_similarity_maps(
-        model,
-        processor,
-        model.device,
-        vit_config,
-        query=query,
-        query_embs=q_embs,
-        token_idx_map=token_to_idx,
-        images=[image_filepath],
-        vespa_sim_maps=None,
-    )
-    for fig_token in figs_images:
-        for token, (fig, ax) in fig_token.items():
-            print(f"Token: {token}")
-            save_figure(fig, f"similarity_map_{token}.png")
-    print("Done")
